fix(module): address Codex feedback for AMD moduleRoot
