+++ crates/swc_ecma_transforms_module/tests/fixture/common/amd-module-root/input.js
@@ -1,0 +1,1 @@
+export const foo = 1;
