+++ crates/swc_ecma_transforms_module/tests/fixture/common/amd-module-root/module.amd.json
@@ -1,0 +1,3 @@
+{
+  "moduleRoot": "tests/fixture/common"
+}
