+++ crates/swc_ecma_transforms_module/tests/fixture/common/amd-module-root/output.amd.js
@@ -1,0 +1,16 @@
+define("amd-module-root/input", [
+    "require",
+    "exports"
+], function(require, exports) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", {
+        value: true
+    });
+    Object.defineProperty(exports, "foo", {
+        enumerable: true,
+        get: function() {
+            return foo;
+        }
+    });
+    const foo = 1;
+});
